+++ sslkeylog.py
@@ -6,12 +6,12 @@ Quickstart::
 
     import sslkeylog
 
-    sslkeylog.set_keylog("sslkeylog.txt")
+    sslkeylog.set_keylog(os.environ.get('SSLKEYLOGFILE'))  # Or directly specify a path
 
     # Do anything involving SSL (Using the built-in ssl module)
 
-And set "(Pre)-Master-Secret log filename" in Wireshark's SSL protocol preferences to the resulting
-file.
+Set the :envvar:`SSLKEYLOGFILE` environment variable if you use it, and set "(Pre)-Master-Secret log
+filename" in Wireshark's SSL protocol preferences to the resulting file.
 """
 
 from __future__ import absolute_import
@@ -97,7 +97,8 @@ def set_keylog(dest):
     """
     global _log_file
 
-    patch()
+    if dest is not None:
+        patch()
 
     if dest is None or callable(dest):
         _sslkeylog._keylog_callback = dest